requirements.txt
asgiref
certifi
channels
channels_redis
charset-normalizer
crispy-bootstrap5
distlib
dj-database-url
Django
django-cors-headers
django-crispy-forms
django-email
django-environ
django-environment
django-filter
django-google-maps
djangorestframework
djangorestframework_simplejwt
filelock
gunicorn
idna
Markdown
MouseInfo
msgpack
packaging
pbr
pillow
pipenv
platformdirs
psycopg
psycopg2
psycopg2-binary
PyAutoGUI
PyGetWindow
PyJWT
PyMsgBox
pyperclip
PyRect
PyScreeze
python-dotenv
pytweening
redis
requests
setuptools
sqlparse
stevedore
typing_extensions
tzdata
urllib3
virtualenv
virtualenv-clone
whitenoise